* some style fixes * fixed casting error in map creator app * added fbow variant to mapcreator

+++ lib-WAI/source/WAISlamTools.cpp
@@ -1126,8 +1126,8 @@ WAIFrame WAISlamTools::createMarkerFrame(std::string       markerFile,
 {
     cv::Mat markerImgGray = cv::imread(markerFile, cv::IMREAD_GRAYSCALE);
 
-    float fyCam = markerCameraIntrinsic.at<float>(1, 1);
-    float cyCam = markerCameraIntrinsic.at<float>(1, 2);
+    float fyCam = (float)markerCameraIntrinsic.at<double>(1, 1);
+    float cyCam = (float)markerCameraIntrinsic.at<double>(1, 2);
     float fov   = (float)(2.0f * atan2(cyCam, fyCam) * 180.0f / M_PI);
 
     float cx = (float)markerImgGray.cols * 0.5f;
